ROC works, log_reg uses dev and test correctly
The model works now - I restructured the function and cleaned it up. Now I call in the train set, create a numerical boolean array (with 1 = > 75th percentile, else 0), and use that array and the train model to create the glm. Once the model's created, I create a numerical boolean array for the test set the same way, then I use predict() to use the glm to create a series of predictors which can then be used to create another boolean array. The test boolean array is the actual data, and the predict() boolean array is the predicted data - I use these values in roc.plot(actual, predicted, threshold = seq(0, 1, 0.01)). Finally, I call the function twice - once for dev and dev TC, then again for test and test TC.