sensitivity/specificity/confusion matrix imperfect
I changed df$tcB to tcB, data=df -tc -tcB (which used the raw tcB rather than the data frame copy). I also changed preds > 0.5 to preds > mean(preds) for ifelse(preds > mean, 1, 0) because none of the tcB values are greater than 0.5, so I adjusted the ifelse classification to fit the variable data. Now the model has an AUC of ~0.645 for the dev set and ~0.659 for the test set - the sensitivity, specificity, false positive, and false negative rates all need work, but they aren't perfect anymore (which means the model judges the actual data and not the data frame slice I took earlier, which is exactly what we need for classification).