fix test set issue
go from subset() (which is column-based) to a working solution (which is row-based), and this works because the model cares about the columns being the same though the row number doesn't matter. this entails using dplyr::mutate to create a new df which has an indicator variable for sum of ik_glaucoma, ak_glaucoma, ... , and if we have an indicator > 0 then we know that specific row must have at least one of the words we care about. we then use dplyr's filter function to limit our rows to indicator variables > 0, and create a test set with all columns and less rows that is our subset for glaucoma words!